requirements.txt
# Web Framework
# Flask requirements are in requirements_flask.txt

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
PyMuPDF==1.23.8

# HTTP Client
httpx==0.26.0

# Data Handling
pandas==2.2.0

# Utilities
python-dotenv==1.0.1
